chore(posters): separate GEWISPosterHandler into Carousel and Base
