+++ src/modules/handlers/screen/poster/gewis/gewis-poster-screen-controller.ts
@@ -1,8 +1,3 @@
-import {
-  BasePosterResponse,
-  BasePosterScreenController,
-  BorrelModeParams,
-} from '../base-poster-screen-controller';
 import HandlerManager from '../../../../root/handler-manager';
 import { Screen } from '../../../../root/entities';
 import { GewisPosterScreenHandler } from '../../index';
@@ -10,21 +5,27 @@ import { Body, Get, Post, Put, Query, Request, Route, Security, Tags } from 'tso
 import { SecurityNames } from '../../../../../helpers/security';
 import { securityGroups } from '../../../../../helpers/security-groups';
 import logger from '../../../../../logger';
-import express, { Request as ExpressRequest } from 'express';
-import { TrainResponse } from '../ns-trains-service';
+import { Request as ExpressRequest } from 'express';
+import NsTrainsService, { TrainResponse } from '../ns-trains-service';
 import GEWISPosterService, { GEWISPhotoAlbumParams } from './gewis-poster-service';
 import OlympicsService from '../olympics-service';
 import { FeatureEnabled } from '../../../../server-settings';
-import { lookup } from 'mime-types';
+import { Controller } from '@tsoa/runtime';
+import { Poster } from '../poster';
 
-interface GewisPosterResponse extends BasePosterResponse {
+export interface BorrelModeParams {
+  enabled: boolean;
+}
+
+export interface PosterResponse {
+  posters: Poster[];
   borrelMode: boolean;
 }
 
-@Route('handler/screen/gewis-poster')
+@Route('handler/screen/poster/carousel')
 @Tags('Handlers')
 @FeatureEnabled('GewisPosterScreenHandler')
-export class GewisPosterScreenController extends BasePosterScreenController {
+export class GewisPosterScreenController extends Controller {
   protected screenHandler: GewisPosterScreenHandler;
 
   constructor() {
@@ -36,66 +37,27 @@ export class GewisPosterScreenController extends BasePosterScreenController {
       )[0] as GewisPosterScreenHandler;
   }
 
-  @Security(SecurityNames.LOCAL, securityGroups.poster.subscriber)
-  @Get('settings')
-  public getPosterCarouselSettings() {
-    return super.getSettings();
-  }
-
-  @Security(SecurityNames.LOCAL, securityGroups.poster.subscriber)
-  @Get('settings/progress-bar-logo')
-  public async getSettingsProgressBarLogo(@Request() request: express.Request) {
-    const logo = await super.getProgressBarLogo();
-    const res = request?.res;
-    if (logo && res) {
-      const mimeType = lookup(logo.name);
-      let contentType: string;
-      if (!mimeType) {
-        contentType = 'application/octet-stream';
-      } else {
-        contentType = mimeType;
-      }
-
-      res.setHeader('Content-Disposition', 'attachment; filename=' + logo.name);
-      res.setHeader('Content-Type', contentType);
-      res.send(logo.data);
-    }
-  }
-
-  @Security(SecurityNames.LOCAL, securityGroups.poster.subscriber)
-  @Get('settings/custom-stylesheet')
-  public async getSettingsProgressBarStylesheet(@Request() request: express.Request) {
-    const stylesheet = await super.getStylesheet();
-    const res = request?.res;
-    if (stylesheet && res) {
-      const mimeType = lookup(stylesheet.name);
-      let contentType: string;
-      if (!mimeType) {
-        contentType = 'application/octet-stream';
-      } else {
-        contentType = mimeType;
-      }
-
-      res.setHeader('Content-Disposition', 'attachment; filename=' + stylesheet.name);
-      res.setHeader('Content-Type', contentType);
-      res.send(stylesheet.data);
-    }
-  }
-
   @Security(SecurityNames.LOCAL, securityGroups.poster.base)
   @Get('')
-  public async getGewisPosters(
-    @Query() alwaysReturnBorrelPosters?: boolean,
-  ): Promise<GewisPosterResponse> {
-    const postersRes = await super.getPosters();
+  public async getPosters(@Query() alwaysReturnBorrelPosters?: boolean): Promise<PosterResponse> {
+    if (!this.screenHandler.posterManager.posters) {
+      try {
+        await this.screenHandler.posterManager.fetchPosters();
+      } catch (e) {
+        logger.error(e);
+      }
+    }
+    const posters = this.screenHandler.posterManager.posters ?? [];
+
     if (alwaysReturnBorrelPosters || this.screenHandler.borrelMode) {
       return {
-        posters: postersRes.posters,
+        posters: posters,
         borrelMode: this.screenHandler.borrelMode,
       };
     }
+
     return {
-      posters: postersRes.posters.filter((p) => !p.borrelMode),
+      posters: posters.filter((p) => !p.borrelMode),
       borrelMode: false,
     };
   }
@@ -103,7 +65,9 @@ export class GewisPosterScreenController extends BasePosterScreenController {
   @Security(SecurityNames.LOCAL, securityGroups.poster.privileged)
   @Post('force-update')
   public async forceUpdateGewisPosters(@Request() req: ExpressRequest): Promise<void> {
-    super.forceUpdatePosters(req);
+    logger.audit(req.user, 'Force fetch posters from source.');
+    await this.screenHandler.posterManager.fetchPosters();
+    this.screenHandler.forceUpdate();
   }
 
   @Security(SecurityNames.LOCAL, securityGroups.poster.base)
@@ -128,7 +92,7 @@ export class GewisPosterScreenController extends BasePosterScreenController {
   @Security(SecurityNames.LOCAL, securityGroups.poster.base)
   @Get('train-departures')
   public async getTrains(): Promise<TrainResponse[]> {
-    return super.getTrains();
+    return new NsTrainsService().getTrains();
   }
 
   @Security(SecurityNames.LOCAL, securityGroups.poster.base)
@@ -140,7 +104,7 @@ export class GewisPosterScreenController extends BasePosterScreenController {
   @Security(SecurityNames.LOCAL, securityGroups.poster.base)
   @Get('olympics/medal-table')
   public async getOlympicsMedalTable() {
-    return super.getOlympicsMedalTable();
+    return new OlympicsService().getMedalTable();
   }
 
   @Security(SecurityNames.LOCAL, securityGroups.poster.base)
